+++ CHANGELOG.md
@@ -6,6 +6,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [**Keep a Changelog v1.0.0**](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [**Semantic Versioning v2.0.0**](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased ##
+
+### Added ###
+
+* New Options:
+  * `--output` - redirect output to another file
+  * `--truncate` - truncate output file
+  * `--append` - append to output file
+
 ## [v0.1.1] - 2021-10-16 ##
 
 [v0.1.1]: https://github.com/mfederczuk/spp/releases/tag/v0.1.1